fix of view controller from storyboard and seperate vc and storyboard for movie list
