+++ MoviesDemoApp/Resources/SceneDelegate.swift
@@ -16,7 +16,13 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
-        guard let _ = (scene as? UIWindowScene) else { return }
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        let storyboard = UIStoryboard(name: "MovieListStoryboard", bundle: nil)
+        let vc = storyboard.instantiateViewController(withIdentifier: "MovieListStoryboard") as! MovieListViewController
+        let navVC = UINavigationController(rootViewController: vc)
+        navVC.navigationBar.prefersLargeTitles = true
+        navVC.viewControllers.first?.navigationItem.largeTitleDisplayMode = .always
+        window?.rootViewController = navVC
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
