+++ MoviesDemoApp/ViewController.swift
@@ -1,0 +1,21 @@
+//
+//  ViewController.swift
+//  MoviesDemoApp
+//
+//  Created by karma on 9/19/22.
+//
+
+//movie list view controller
+
+import UIKit
+
+class ViewController: UIViewController {
+    
+    // MARK: - Lifecycles
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+    }
+    
+}
+
